// src/app/models/ConfigResource.model.ts
export class ConfigResource {
  title: string;

  mail: string;
  site: string;
  github: string;

  youtubeChannelId: string;
  facebookPageId: string;
  facebookGroupId: string;
  twitter: string;

  blog: string;
  blogFeed: string;

  googlecalendar: string;
  googlegroup: string;
  flickr: string;
}